refactor: Seperate out MapView into mapview pkg; rename map pkg to tiles
Allows simpler integration / usage as a library since MapView logic can
be reused. Also adds a few minor tooling fixups.
